Update tests: remove general security test in favor of permission checker
