+++ nuget-feed/src/jetbrains/buildServer/nuget/feed/server/index/impl/security/NuGetFeedPermissionChecker.kt
@@ -4,6 +4,9 @@ import jetbrains.buildServer.nuget.feed.server.index.NuGetFeedData
 import jetbrains.buildServer.serverSide.SBuild
 import jetbrains.buildServer.serverSide.SProject
 
+/**
+ * Authorizes NuGet package publishing: a project may write only to feeds available to it.
+ */
 interface NuGetFeedPermissionChecker {
     fun canWrite(build: SBuild, feed: NuGetFeedData): Boolean
