+++ nuget-tests/src/testng-nuget-fast.xml
@@ -19,7 +19,7 @@
       <class name="jetbrains.buildServer.nuget.tests.server.feed.server.RequestWrapperTest"/>
       <class name="jetbrains.buildServer.nuget.tests.server.feed.server.NuGetArtifactsMetadataProviderTest"/>
       <class name="jetbrains.buildServer.nuget.tests.server.feed.server.NuGetBuildFeedsProviderTest"/>
-      <class name="jetbrains.buildServer.nuget.tests.server.feed.server.NuGetFeedProviderSecurityTest"/>
+      <class name="jetbrains.buildServer.nuget.tests.server.feed.server.security.NuGetFeedPermissionCheckerTest"/>
       <class name="jetbrains.buildServer.nuget.tests.server.feed.server.NuGetBuildMetadataProviderTest"/>
       <class name="jetbrains.buildServer.nuget.tests.server.SemanticVersionTest"/>
       <class name="jetbrains.buildServer.nuget.tests.server.SemanticVersionComparisonTest"/>
